ocr_automation/requirements.txt
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.1,<2.0
google-api-python-client>=2.120.0,<3.0
google-auth>=2.28.0,<3.0
google-auth-httplib2>=0.2.0,<1.0
requests>=2.31.0,<3.0
pytest>=7.4.0,<9.0
